+++ backend/interface/ch.ini
@@ -6,18 +6,93 @@ Mode = 选择识别模式:
 ModeFast = 快速
 ModeAccurate = 精准
 InterfaceDefault = 简体中文
-LanguageSimplifiedChinese = 简体中文
-LanguageTraditionalChinese = 繁体中文
-LanguageEnglish = 英语
-LanguageJapanese = 日语
-LanguageKorean = 韩语
-LanguageArabic = 阿拉伯语
-LanguageFrench = 法语
-LanguageGerman = 德语
-LanguageRussian = 俄语
-LanguageSpanish = 西班牙语
-LanguagePortuguese = 葡萄牙语
-LanguageItalian = 意大利语
+LanguageCH = 简体中文
+LanguageCHINESE_CHT = 繁体中文
+LanguageEN = 英文
+LanguageJAPAN = 日文
+LanguageKOREAN = 韩文
+LanguageAR = 阿拉伯文
+LanguageFRENCH = 法文
+LanguageGERMAN = 德文
+LanguageRU = 俄罗斯文
+LanguageES = 西班牙文
+LanguagePT = 葡萄牙文
+LanguageIT = 意大利文
+LanguageAF = 南非荷兰文
+LanguageAZ = 阿塞拜疆文
+LanguageBS = 波斯尼亚文
+LanguageCS = 捷克文
+LanguageCY = 威尔士文
+LanguageDA = 丹麦文
+LanguageDE = 德文
+LanguageET = 爱沙尼亚文
+LanguageFR = 法文
+LanguageGA = 爱尔兰文
+LanguageHR = 克罗地亚文
+LanguageHU = 匈牙利文
+LanguageID = 印尼文
+LanguageIS = 冰岛文
+LanguageKU = 库尔德文
+LanguageLA = 拉丁文
+LanguageLT = 立陶宛文
+LanguageLV = 拉脱维亚文
+LanguageMI = 毛利文
+LanguageMS = 马来文
+LanguageMT = 马耳他文
+LanguageNL = 荷兰文
+LanguageNO = 挪威文
+LanguageOC = 欧西坦文
+LanguagePI = 巴利文
+LanguagePL = 波兰文
+LanguageRO = 罗马尼亚文
+LanguageRS_LATIN = 塞尔维亚文(latin)
+LanguageSK = 斯洛伐克文
+LanguageSL = 斯洛文尼亚文
+LanguageSQ = 阿尔巴尼亚文
+LanguageSV = 瑞典文
+LanguageSW = 西瓦希里文
+LanguageTL = 塔加洛文
+LanguageTR = 土耳其文
+LanguageUZ = 乌兹别克文
+LanguageVI = 越南文
+LanguageLATIN = 拉丁文
+LanguageFA = 波斯文
+LanguageUG = 维吾尔文
+LanguageUR = 乌尔都文
+LanguageRS_CYRILLIC = 塞尔维亚文(cyrillic)
+LanguageBE = 白俄罗斯文
+LanguageBG = 保加利亚文
+LanguageUK = 乌克兰文
+LanguageMN = 蒙古文
+LanguageABQ = 阿巴扎文
+LanguageADY = 阿迪赫文
+LanguageKBD = 卡巴尔达文
+LanguageAVA = 阿瓦尔文
+LanguageDAR = 达尔瓦文
+LanguageINH = 因古什文
+LanguageCHE = 车臣文
+LanguageLBE = 拉克文
+LanguageLEZ = 莱兹甘文
+LanguageTAB = 塔巴萨兰文
+LanguageCYRILLIC = 西里尔文
+LanguageHI = 印地文
+LanguageMR = 马拉地文
+LanguageNE = 尼泊尔文
+LanguageBH = 比尔哈文
+LanguageMAI = 迈蒂利文
+LanguageANG = 昂加文
+LanguageBHO = 孟加拉文
+LanguageMAH = 摩揭陀文
+LanguageSCK = 那格浦尔文
+LanguageNEW = 尼瓦尔文
+LanguageGOM = 保加利亚文
+LanguageSA = 沙特阿拉伯文
+LanguageBGC = 哈里亚纳文
+LanguageDEVANAGARI = 德瓦那加里文
+LanguageTA = 泰米尔文
+LanguageKN = 卡纳达文
+LanguageTE = 泰卢固文
+LanguageKA = 卡纳达文
 
 [SubtitleExtractorGUI]
 Title = 字幕提取器
