+++ backend/interface/chinese_cht.ini
@@ -6,18 +6,93 @@ Mode = 選擇識別模式:
 ModeFast = 快速
 ModeAccurate = 精準
 InterfaceDefault = 繁體中文
-LanguageSimplifiedChinese = 簡體中文
-LanguageTraditionalChinese = 繁體中文
-LanguageEnglish = 英語
-LanguageJapanese = 日語
-LanguageKorean = 韓語
-LanguageArabic = 阿拉伯語
-LanguageFrench = 法語
-LanguageGerman = 德語
-LanguageRussian = 俄語
-LanguageSpanish = 西班牙語
-LanguagePortuguese = 葡萄牙語
-LanguageItalian = 意大利語
+LanguageCH = 簡體中文
+LanguageCHINESE_CHT = 繁體中文
+LanguageEN = 英文
+LanguageJAPAN = 日文
+LanguageKOREAN = 韓文
+LanguageAR = 阿拉伯文
+LanguageFRENCH = 法文
+LanguageGERMAN = 德文
+LanguageRU = 俄羅斯文
+LanguageES = 西班牙文
+LanguagePT = 葡萄牙文
+LanguageIT = 意大利文
+LanguageAF = 南非荷蘭文
+LanguageAZ = 阿塞拜疆文
+LanguageBS = 波斯尼亞文
+LanguageCS = 捷克文
+LanguageCY = 威爾士文
+LanguageDA = 丹麥文
+LanguageDE = 德文
+LanguageET = 愛沙尼亞文
+LanguageFR = 法文
+LanguageGA = 愛爾蘭文
+LanguageHR = 克羅地亞文
+LanguageHU = 匈牙利文
+LanguageID = 印尼文
+LanguageIS = 冰島文
+LanguageKU = 庫爾德文
+LanguageLA = 拉丁文
+LanguageLT = 立陶宛文
+LanguageLV = 拉脫維亞文
+LanguageMI = 毛利文
+LanguageMS = 馬來文
+LanguageMT = 馬耳他文
+LanguageNL = 荷蘭文
+LanguageNO = 挪威文
+LanguageOC = 歐西坦文
+LanguagePI = 巴利文
+LanguagePL = 波蘭文
+LanguageRO = 羅馬尼亞文
+LanguageRS_LATIN = 塞爾維亞文(latin)
+LanguageSK = 斯洛伐克文
+LanguageSL = 斯洛文尼亞文
+LanguageSQ = 阿爾巴尼亞文
+LanguageSV = 瑞典文
+LanguageSW = 西瓦希裏文
+LanguageTL = 塔加洛文
+LanguageTR = 土耳其文
+LanguageUZ = 烏茲別克文
+LanguageVI = 越南文
+LanguageLATIN = 拉丁文
+LanguageFA = 波斯文
+LanguageUG = 維吾爾文
+LanguageUR = 烏爾都文
+LanguageRS_CYRILLIC = 塞爾維亞文(cyrillic)
+LanguageBE = 白俄羅斯文
+LanguageBG = 保加利亞文
+LanguageUK = 烏克蘭文
+LanguageMN = 蒙古文
+LanguageABQ = 阿巴紮文
+LanguageADY = 阿迪赫文
+LanguageKBD = 卡巴爾達文
+LanguageAVA = 阿瓦爾文
+LanguageDAR = 達爾瓦文
+LanguageINH = 因古什文
+LanguageCHE = 車臣文
+LanguageLBE = 拉克文
+LanguageLEZ = 萊茲甘文
+LanguageTAB = 塔巴薩蘭文
+LanguageCYRILLIC = 西裏爾文
+LanguageHI = 印地文
+LanguageMR = 馬拉地文
+LanguageNE = 尼泊爾文
+LanguageBH = 比爾哈文
+LanguageMAI = 邁蒂利文
+LanguageANG = 昂加文
+LanguageBHO = 孟加拉文
+LanguageMAH = 摩揭陀文
+LanguageSCK = 那格浦爾文
+LanguageNEW = 尼瓦爾文
+LanguageGOM = 保加利亞文
+LanguageSA = 沙特阿拉伯文
+LanguageBGC = 哈裏亞納文
+LanguageDEVANAGARI = 德瓦那加裏文
+LanguageTA = 泰米爾文
+LanguageKN = 卡納達文
+LanguageTE = 泰盧固文
+LanguageKA = 卡納達文
 
 [SubtitleExtractorGUI]
 Title = 字幕提取器
