+++ backend/interface/en.ini
@@ -6,18 +6,93 @@ Mode = Choose Mode:
 ModeFast = fast
 ModeAccurate = accurate
 InterfaceDefault = English
-LanguageSimplifiedChinese = Simplified Chinese
-LanguageTraditionalChinese = Traditional Chinese
-LanguageEnglish = English
-LanguageJapanese = Japanese
-LanguageKorean = Korean
-LanguageArabic = Arabic
-LanguageFrench = French
-LanguageGerman = German
-LanguageRussian = Russian
-LanguageSpanish = Spanish
-LanguagePortuguese = Portuguese
-LanguageItalian = Italian
+LanguageCH = Simplified Chinese
+LanguageCHINESE_CHT = Traditional Chinese
+LanguageEN = English
+LanguageJAPAN = Japanese
+LanguageKOREAN = Korean
+LanguageAR = Arabic
+LanguageFRENCH = French
+LanguageGERMAN = German
+LanguageRU = Russian
+LanguageES = Spanish
+LanguagePT = Portuguese
+LanguageIT = Italian
+LanguageAF = Afrikaans
+LanguageAZ = Azerbaijani
+LanguageBS = Bosnian
+LanguageCS = Czech
+LanguageCY = Welsh
+LanguageDA = Danish
+LanguageDE = German
+LanguageET = Estonian
+LanguageFR = French
+LanguageGA = Irish
+LanguageHR = Croatian
+LanguageHU = Hungarian
+LanguageID = Indonesian
+LanguageIS = Icelandic
+LanguageKU = Kurdish
+LanguageLA = Latin
+LanguageLT = Lithuanian
+LanguageLV = Latvian
+LanguageMI = Maori
+LanguageMS = Malay
+LanguageMT = Maltese
+LanguageNL = Dutch
+LanguageNO = Norwegian
+LanguageOC = Occitan
+LanguagePI = Pali
+LanguagePL = Polish
+LanguageRO = Romanian
+LanguageRS_LATIN = Serbian(latin)
+LanguageSK = Slovak
+LanguageSL = Slovenian
+LanguageSQ = Albanian
+LanguageSV = Swedish
+LanguageSW = Swahili
+LanguageTL = Tagalog
+LanguageTR = Turkish
+LanguageUZ = Uzbek
+LanguageVI = Vietnamese
+LanguageLATIN = Latin
+LanguageFA = Persian
+LanguageUR = Urdu
+LanguageRS_CYRILLIC = Serbian(cyrillic)
+LanguageBE = Belarusian
+LanguageBG = Bulgarian
+LanguageUK = Ukranian
+LanguageMN = Mongolian
+LanguageABQ = Abaza
+LanguageADY = Adyghe
+LanguageKBD = Kabardian
+LanguageAVA = Avar
+LanguageDAR = Dargwa
+LanguageINH = Ingush
+LanguageCHE = Chechen
+LanguageLBE = Lak
+LanguageLEZ = Lezghian
+LanguageTAB = Tabassaran
+LanguageCYRILLIC = Cyrillic
+LanguageHI = Hindi
+LanguageMR = Marathi
+LanguageNE = Nepali
+LanguageBH = Bihari
+LanguageMAI = Maithili
+LanguageANG = Angika
+LanguageBHO = Bhojpuri
+LanguageMAH = Magahi
+LanguageSCK = Nagpur
+LanguageNEW = Newari
+LanguageGOM = Goan Konkani
+LanguageSA = Saudi Arabia
+LanguageBGC = Haryanvi
+LanguageDEVANAGARI = Devanagari
+LanguageTA = Tamil
+LanguageKN = Kannada
+LanguageUG = Uyghur
+LanguageTE = Telugu
+LanguageKA = Kannada
 
 [SubtitleExtractorGUI]
 Title = Subtitle Extractor
