+++ backend/ppocr/utils/dict/ar_dict.txt
@@ -1,162 +1,117 @@
- 
-!
-#
-$
-%
-&
-'
-(
-+
-,
--
-.
-/
-0
-1
-2
-3
-4
-5
-6
-7
-8
-9
-:
-?
-@
-A
-B
-C
-D
-E
-F
-G
-H
-I
-J
-K
-L
-M
-N
-O
-P
-Q
-R
-S
-T
-U
-V
-W
-X
-Y
-Z
-_
 a
-b
-c
-d
-e
-f
-g
-h
-i
-j
-k
-l
-m
-n
-o
-p
-q
 r
-s
-t
-u
+b
+i
+c
+_
+m
+g
+/
+1
+0
+I
+L
+S
+V
+R
+C
+2
 v
-w
-x
-y
-z
-É
-é
-ء
-آ
-أ
-ؤ
-إ
-ئ
+l
+6
+3
+9
+.
+j
+p	
 ا
-ب
-ة
-ت
-ث
-ج
-ح
-خ
-د
-ذ
-ر
-ز
-س
-ش
-ص
-ض
-ط
-ظ
-ع
-غ
-ف
-ق
-ك
 ل
 م
-ن
-ه
+ر
+ج
 و
-ى
+ح
 ي
-ً
-ٌ
-ٍ
-َ
-ُ
-ِ
-ّ
-ْ
-ٓ
-ٔ
-ٰ
-ٱ
-ٹ
-پ
-چ
-ڈ
-ڑ
-ژ
-ک
-ڭ
-گ
-ں
-ھ
-ۀ
-ہ
-ۂ
-ۃ
-ۆ
-ۇ
-ۈ
-ۋ
-ی
-ې
-ے
-ۓ
-ە
-١
-٢
-٣
-٤
-٥
-٦
-٧
-٨
-٩
+ة
+5
+8
+7
+أ
+ب
+ض
+4
+ك
+س
+ه
+ث
+ن
+ط
+ع
+ت
+غ
+خ
+ف
+ئ
+ز
+إ
+د
+ص
+ظ
+ذ
+ش
+ى
+ق
+ؤ
+آ
+ء
+s
+e
+n
+w
+t
+u
+z
+d
+A
+N
+G
+h
+o
+E
+T
+H
+O
+B
+y
+F
+U
+J
+X
+W
+P
+Z
+M
+k
+q
+Y
+Q
+D
+f
+K
+x
+'
+%
+-
+#
+@
+!
+&
+$
+,
+:
+é
+?
++
+É
+(
+
