+++ backend/ppocr/utils/dict/kie_dict/xfund_class_list.txt
@@ -1,0 +1,4 @@
+OTHER
+QUESTION
+ANSWER
+HEADER
