+++ backend/ppocr/utils/dict/layout_dict/layout_cdla_dict.txt
@@ -1,0 +1,10 @@
+text
+title
+figure
+figure_caption
+table
+table_caption
+header
+footer
+reference
+equation
