+++ backend/ppocr/utils/dict/layout_dict/layout_publaynet_dict.txt
@@ -1,0 +1,5 @@
+text
+title
+list
+table
+figure
