+++ backend/ppocr/utils/dict/layout_dict/layout_table_dict.txt
@@ -1,0 +1,1 @@
+table
